rename ClassReplacerRector to RenameClassRector
