+++ config/level/cakephp/cakephp35.yaml
@@ -1,7 +1,7 @@
 # source: https://book.cakephp.org/3.0/en/appendices/3-5-migration-guide.html
 
 services:
-    Rector\Rector\Class_\ClassReplacerRector:
+    Rector\Rector\Class_\RenameClassRector:
         Cake\Http\Client\CookieCollection: 'Cake\Http\Cookie\CookieCollection'
         Cake\Console\ShellDispatcher: 'Cake\Console\CommandRunner'
