+++ config/level/php-code-sniffer/php-code-sniffer30.yaml
@@ -1,5 +1,5 @@
 services:
-    Rector\Rector\Class_\ClassReplacerRector:
+    Rector\Rector\Class_\RenameClassRector:
         'PHP_CodeSniffer_Sniffs_Sniff': 'PHP_CodeSniffer\Sniffs\Sniff'
         'PHP_CodeSniffer_File': 'PHP_CodeSniffer\Files\File'
         'PHP_CodeSniffer_Tokens': 'PHP_CodeSniffer\Util\Tokens'
