+++ config/level/symfony/symfony40.yaml
@@ -8,6 +8,6 @@ services:
     Rector\Symfony\Rector\Process\ProcessBuilderInstanceRector: ~
     Rector\Symfony\Rector\Process\ProcessBuilderGetProcessRector: ~
 
-    Rector\Rector\Class_\ClassReplacerRector:
+    Rector\Rector\Class_\RenameClassRector:
         'Symfony\Component\Validator\Tests\Constraints\AbstractConstraintValidatorTest': 'Symfony\Component\Validator\Test\ConstraintValidatorTestCase'
         'Symfony\Component\Process\ProcessBuilder': 'Symfony\Component\Process\Process'
