+++ config/level/twig/twig112.yaml
@@ -1,6 +1,6 @@
 services:
     # see https://stackoverflow.com/a/43495506/1348344
-    Rector\Rector\Class_\ClassReplacerRector:
+    Rector\Rector\Class_\RenameClassRector:
         Twig_Function_Node: 'Twig_SimpleFunction'
         Twig_Function: 'Twig_SimpleFunction'
         Twig_Filter: 'Twig_SimpleFilter'
