--- tests/Rector/Class_/ClassReplacerRector/Fixture/class_to_new.php.inc
@@ -1,21 +1,0 @@
-<?php
-
-namespace Rector\Tests\Rector\Class_\ClassReplacerRector;
-
-class MyCustomValidatorTest extends \Rector\Tests\Rector\Class_\ClassReplacerRector\Source\OldClass
-{
-    // ...
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\Rector\Class_\ClassReplacerRector;
-
-class MyCustomValidatorTest extends \Rector\Tests\Rector\Class_\ClassReplacerRector\Source\NewClass
-{
-    // ...
-}
-
-?>
