--- tests/Rector/Class_/ClassReplacerRector/Source/NewClassWithoutTypo.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\Class_\ClassReplacerRector\Source;
-
-class NewClassWithoutTypo
-{
-
-}
