+++ tests/Rector/Class_/RenameClassRector/Fixture/class_to_new.php.inc
@@ -1,0 +1,21 @@
+<?php
+
+namespace Rector\Tests\Rector\Class_\RenameClassRector;
+
+class MyCustomValidatorTest extends \Rector\Tests\Rector\Class_\RenameClassRector\Source\OldClass
+{
+    // ...
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\Rector\Class_\RenameClassRector;
+
+class MyCustomValidatorTest extends \Rector\Tests\Rector\Class_\RenameClassRector\Source\NewClass
+{
+    // ...
+}
+
+?>
