+++ tests/Rector/Class_/RenameClassRector/Fixture/keep_return_tag.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Rector\Class_\ClassReplacerRector\Fixture;
+namespace Rector\Tests\Rector\Class_\RenameClassRector\Fixture;
 
 use Manual\Twig\TwigFilter;
