+++ tests/Rector/Class_/RenameClassRector/Fixture/name_insensitive.php.inc
@@ -1,0 +1,31 @@
+<?php
+
+namespace Rector\Tests\Rector\Class_\RenameClassRector\Fixture;
+
+use Rector\Tests\Rector\Class_\RenameClassRector\Source\OldClassWithTypo;
+
+class NameInsensitive extends OldClassWithTypO
+{
+    public function run(): OLDClassWithTYPO
+    {
+        $oldClassWithTypo = new OldClassWithTYPO;
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\Rector\Class_\RenameClassRector\Fixture;
+
+use \Rector\Tests\Rector\Class_\RenameClassRector\Source\NewClassWithoutTypo;
+
+class NameInsensitive extends \Rector\Tests\Rector\Class_\RenameClassRector\Source\NewClassWithoutTypo
+{
+    public function run(): \Rector\Tests\Rector\Class_\RenameClassRector\Source\NewClassWithoutTypo
+    {
+        $oldClassWithTypo = new \Rector\Tests\Rector\Class_\RenameClassRector\Source\NewClassWithoutTypo;
+    }
+}
+
+?>
