+++ tests/Rector/Class_/RenameClassRector/Source/AbstractManualExtension.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Class_\ClassReplacerRector\Source;
+namespace Rector\Tests\Rector\Class_\RenameClassRector\Source;
 
 abstract class AbstractManualExtension
 {
