+++ tests/Rector/Class_/RenameClassRector/Source/Manual_Twig_Filter.php
