+++ tests/Rector/Class_/RenameClassRector/Source/OldClass.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\Class_\RenameClassRector\Source;
+
+class OldClass
+{
+
+}
