+++ tests/Rector/Class_/RenameClassRector/Source/TwigFilter.php
